refactor: Extract formLogic for cleaner code
